% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binRep.R
\name{binRep}
\alias{binRep}
\title{Transform DNAStringSet to binary representation
Transforms DNAStringSet to binary representation}
\usage{
binRep(x)
}
\arguments{
\item{x}{sequences as DNAStringSet}
}
\value{
List of sequences in binary represenation.
}
\description{
Transform DNAStringSet to binary representation
Transforms DNAStringSet to binary representation
}
\examples{
binRep(sequences)
}
