% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHashTable.R
\name{buildHashTable}
\alias{buildHashTable}
\title{Build hash dable}
\usage{
buildHashTable(fasta_soubor, k = 10)
}
\arguments{
\item{fasta_soubor}{FASTA file containing database}

\item{k}{number of bases the database is divided into}
}
\value{
List containing list of pointers, list of indexes and list of sequence headers
}
\description{
Builds hash table from FASTA file containing database.
}
\examples{
buildHashTable("database.fasta", k=10)

}
