% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense.R
\name{condense}
\alias{condense}
\title{Demultiplexed data condensing}
\usage{
condense(fasta_soubor, prah = 10)
}
\arguments{
\item{fasta_soubor}{file demultiplexed data}

\item{prah}{cutoff for number of unique sequences, sequences with amount less than
the cutoff are not saved in the final FASTA file}
}
\value{
None
}
\description{
Proccesses demultiplexed fasta files, finds identical reads and deletes them.
The number of identical reads is retained in sequence headers.
Condensed reads are saved as FASTA file.
}
\examples{
condense("demultiplexed_reads.fasta", prah=10)

}
