% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert binary to int}
\usage{
convert(s)
}
\arguments{
\item{s}{set of binary numbers}
}
\value{
None
}
\description{
Tranforms set of binary numbers to integers.
}
\examples{
convert(set)

}
