% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex}
\alias{demultiplex}
\title{Raw data demultiplexing}
\usage{
demultiplex(fasta_soubor, f_primer, r_primer)
}
\arguments{
\item{fasta_soubor}{file with raw data in FASTA format}

\item{f_primer}{forward MID}

\item{r_primer}{reverse MID}
}
\value{
None
}
\description{
User inputs fasta file and MIDs an function outputs a fasta file of demultiplexed reads.
Demultiplexed reads are saved as FASTA.
}
\examples{
demultiplex("MHC_reads.fasta", "ACTGACTGGT", "ATGACGTAGT")

}
