% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplexAll.R
\name{demultiplex_all}
\alias{demultiplex_all}
\title{Continuous raw data demultiplexing}
\usage{
demultiplex_all(fasta_soubor)
}
\arguments{
\item{fasta_soubor}{file with raw data in FASTA format}
}
\value{
None
}
\description{
User inputs fasta file and the function then asks him to input MIDs for various reads.
Demultiplexed reads are saved as FASTA.
}
\examples{
demultiplex_all("MHC_reads.fasta")

}
