% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDb.R
\name{downloadDb}
\alias{downloadDb}
\title{Database download}
\usage{
downloadDb(URL, type = "nuc", directory = getwd(),
  mergeIntoOneFasta = FALSE, name = "database")
}
\arguments{
\item{URL}{URL to database on EBI}

\item{type}{type of sequences (nuc for nucletide, prot for aminoacid)}

\item{directory}{directory to which the database is saved}

\item{mergeIntoOneFasta}{TRUE: merge all FASTA files to a single FASTA file}

\item{name}{name of the merged FASTA file}
}
\value{
None
}
\description{
Downloads database from EBI ftp server.
}
\examples{
download_DB('ftp://ftp.ebi.ac.uk/pub/databases/ipd/mhc/nhp/')

}
