% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchHashTable.R
\name{searchHashTable}
\alias{searchHashTable}
\title{Search hash table}
\usage{
searchHashTable(querySequences, hashTable, k = 10)
}
\arguments{
\item{querySequences}{FASTA file containing sequences we are searching for}

\item{hashTable}{output of buildHashTable function}

\item{k}{number of bases the query sequences are divided into, must be the same as in buildHashTable}
}
\value{
A matrix containg best match and score for each sequence
}
\description{
Searches hash table form buildHashTable for query sequences from FASTA file.
}
\examples{
searchHashTable("query.fasta", hashTable k=10)

}
