% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchHashTable.R
\name{searchHashTable}
\alias{searchHashTable}
\title{Search hash table}
\usage{
searchHashTable(querySequences, hashTable, k = 10, results = 1,
  tolerance = 1)
}
\arguments{
\item{querySequences}{FASTA file containing sequences we are searching for}

\item{hashTable}{output of buildHashTable function}

\item{k}{number of bases the query sequences are divided into, must be the same as in buildHashTable}

\item{results}{number of top results of the SSAHA method that will be provie in the outpus list}

\item{tolerance}{tolerance for insertions and deletions in number of nucleotides}
}
\value{
A list of matrices containg best matches and score for each query sequence.
}
\description{
Searches hash table obtained using the buildHashTable function for occurences of query sequences from FASTA file.
}
\examples{
searchHashTable("query.fasta", hashTable, k=10, results = 10, tolerance = 5)

}
